--- agucareer/.packages
@@ -1,45 +1,0 @@
-# Generated by pub on 2020-05-13 17:19:05.283588.
-archive:file:///C:/src/flutter/.pub-cache/hosted/pub.dartlang.org/archive-2.0.11/lib/
-args:file:///C:/src/flutter/.pub-cache/hosted/pub.dartlang.org/args-1.5.2/lib/
-async:file:///C:/src/flutter/.pub-cache/hosted/pub.dartlang.org/async-2.4.0/lib/
-boolean_selector:file:///C:/src/flutter/.pub-cache/hosted/pub.dartlang.org/boolean_selector-1.0.5/lib/
-charcode:file:///C:/src/flutter/.pub-cache/hosted/pub.dartlang.org/charcode-1.1.2/lib/
-cloud_firestore:file:///C:/src/flutter/.pub-cache/hosted/pub.dartlang.org/cloud_firestore-0.13.5/lib/
-cloud_firestore_platform_interface:file:///C:/src/flutter/.pub-cache/hosted/pub.dartlang.org/cloud_firestore_platform_interface-1.1.0/lib/
-cloud_firestore_web:file:///C:/src/flutter/.pub-cache/hosted/pub.dartlang.org/cloud_firestore_web-0.1.1+2/lib/
-collection:file:///C:/src/flutter/.pub-cache/hosted/pub.dartlang.org/collection-1.14.11/lib/
-convert:file:///C:/src/flutter/.pub-cache/hosted/pub.dartlang.org/convert-2.1.1/lib/
-crypto:file:///C:/src/flutter/.pub-cache/hosted/pub.dartlang.org/crypto-2.1.3/lib/
-cupertino_icons:file:///C:/src/flutter/.pub-cache/hosted/pub.dartlang.org/cupertino_icons-0.1.3/lib/
-firebase:file:///C:/src/flutter/.pub-cache/hosted/pub.dartlang.org/firebase-7.3.0/lib/
-firebase_auth:file:///C:/src/flutter/.pub-cache/hosted/pub.dartlang.org/firebase_auth-0.16.0/lib/
-firebase_auth_platform_interface:file:///C:/src/flutter/.pub-cache/hosted/pub.dartlang.org/firebase_auth_platform_interface-1.1.7/lib/
-firebase_auth_web:file:///C:/src/flutter/.pub-cache/hosted/pub.dartlang.org/firebase_auth_web-0.1.2/lib/
-firebase_core:file:///C:/src/flutter/.pub-cache/hosted/pub.dartlang.org/firebase_core-0.4.4+3/lib/
-firebase_core_platform_interface:file:///C:/src/flutter/.pub-cache/hosted/pub.dartlang.org/firebase_core_platform_interface-1.0.4/lib/
-firebase_core_web:file:///C:/src/flutter/.pub-cache/hosted/pub.dartlang.org/firebase_core_web-0.1.1+2/lib/
-flutter:file:///C:/src/flutter/packages/flutter/lib/
-flutter_test:file:///C:/src/flutter/packages/flutter_test/lib/
-flutter_web_plugins:file:///C:/src/flutter/packages/flutter_web_plugins/lib/
-http:file:///C:/src/flutter/.pub-cache/hosted/pub.dartlang.org/http-0.12.1/lib/
-http_parser:file:///C:/src/flutter/.pub-cache/hosted/pub.dartlang.org/http_parser-3.1.4/lib/
-image:file:///C:/src/flutter/.pub-cache/hosted/pub.dartlang.org/image-2.1.4/lib/
-js:file:///C:/src/flutter/.pub-cache/hosted/pub.dartlang.org/js-0.6.1+1/lib/
-matcher:file:///C:/src/flutter/.pub-cache/hosted/pub.dartlang.org/matcher-0.12.6/lib/
-meta:file:///C:/src/flutter/.pub-cache/hosted/pub.dartlang.org/meta-1.1.8/lib/
-path:file:///C:/src/flutter/.pub-cache/hosted/pub.dartlang.org/path-1.6.4/lib/
-pedantic:file:///C:/src/flutter/.pub-cache/hosted/pub.dartlang.org/pedantic-1.8.0+1/lib/
-petitparser:file:///C:/src/flutter/.pub-cache/hosted/pub.dartlang.org/petitparser-2.4.0/lib/
-plugin_platform_interface:file:///C:/src/flutter/.pub-cache/hosted/pub.dartlang.org/plugin_platform_interface-1.0.2/lib/
-quiver:file:///C:/src/flutter/.pub-cache/hosted/pub.dartlang.org/quiver-2.0.5/lib/
-sky_engine:file:///C:/src/flutter/bin/cache/pkg/sky_engine/lib/
-source_span:file:///C:/src/flutter/.pub-cache/hosted/pub.dartlang.org/source_span-1.5.5/lib/
-stack_trace:file:///C:/src/flutter/.pub-cache/hosted/pub.dartlang.org/stack_trace-1.9.3/lib/
-stream_channel:file:///C:/src/flutter/.pub-cache/hosted/pub.dartlang.org/stream_channel-2.0.0/lib/
-string_scanner:file:///C:/src/flutter/.pub-cache/hosted/pub.dartlang.org/string_scanner-1.0.5/lib/
-term_glyph:file:///C:/src/flutter/.pub-cache/hosted/pub.dartlang.org/term_glyph-1.1.0/lib/
-test_api:file:///C:/src/flutter/.pub-cache/hosted/pub.dartlang.org/test_api-0.2.11/lib/
-typed_data:file:///C:/src/flutter/.pub-cache/hosted/pub.dartlang.org/typed_data-1.1.6/lib/
-vector_math:file:///C:/src/flutter/.pub-cache/hosted/pub.dartlang.org/vector_math-2.0.8/lib/
-xml:file:///C:/src/flutter/.pub-cache/hosted/pub.dartlang.org/xml-3.5.0/lib/
-agucareer:lib/
