+++ agucareer/lib/locator.dart
@@ -1,0 +1,10 @@
+import 'package:agucareer/repository/user_repository.dart';
+import 'package:agucareer/services/firebase_auth_service.dart';
+import 'package:get_it/get_it.dart';
+
+GetIt locator = GetIt.instance;
+
+void setupLocator(){
+  locator.registerLazySingleton(() => FirebaseAuthService());
+  locator.registerLazySingleton(() => UserRepository());
+}
