+++ agucareer/lib/main.dart
@@ -1,16 +1,23 @@
-import 'package:agucareer/hosgeldin_widget/hosgeldin_widget.dart';
+import 'package:agucareer/locator.dart';
+import 'package:agucareer/pages/landing_page.dart';
+import 'package:agucareer/pages/login_page.dart';
+import 'package:agucareer/viewmodels/user_model.dart';
 import 'package:flutter/material.dart';
+import 'package:provider/provider.dart';
 
-void main() => runApp(MyApp());
-
+void main() {
+  setupLocator();
+  runApp(MyApp());
+}
 
 class MyApp extends StatelessWidget {
-
   @override
   Widget build(BuildContext context) {
-
-    return MaterialApp(
-      home: HosgeldinWidget(),
-    );
+    return ChangeNotifierProvider(
+        create: (context) => UserModel(),
+        child: MaterialApp(
+          debugShowCheckedModeBanner: false,
+          home: LandingPage(),
+        ));
   }
 }
