+++ agucareer/lib/models/user_model.dart
@@ -1,0 +1,7 @@
+import 'package:flutter/cupertino.dart';
+
+class User {
+  final String userID;
+
+  User({@required this.userID});
+}
