+++ agucareer/lib/values/colors.dart
@@ -2,6 +2,12 @@ import 'dart:ui';
 
 
 class AppColors {
+  static const Color koyuMor = Color.fromRGBO(36, 19, 50, 100);
+  static const Color acikMor = Color.fromRGBO(138, 86, 172, 100);
+  static const Color pembe = Color.fromRGBO(212 , 127, 166, 100);
+  static const Color acikMavi = Color.fromRGBO(46, 165, 255, 100);
+  static const Color yaziGri = Color.fromRGBO(112, 112, 112, 56);
+
   static const Color mainBackground = Color.fromARGB(255, 121, 199, 212);
   static const Color mainSecondaryBackground = Color.fromARGB(0, 121, 199, 212);
   static const Color splashButtonColor = Color.fromARGB(175, 18, 68, 64);
