+++ agucareer/pubspec.yaml
@@ -21,6 +21,8 @@ dependencies:
     sdk: flutter
   firebase_auth: ^0.16.0 #for firebase
   cloud_firestore: ^0.13.5 #for firestore
+  get_it: ^4.0.2
+  provider: ^4.0.5+1
 
   # The following adds the Cupertino Icons font to your application.
   # Use with the CupertinoIcons class for iOS style icons.
